requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
groq==0.9.0
sqlalchemy==2.0.30
python-multipart==0.0.9